refactoring demography and logging
