+++ Protecs/src/test/java/uk/ac/ucl/protecs/behaviours/MobilityTesting.java
@@ -30,4 +30,5 @@ public class MobilityTesting {
 		WorldBankCovid19Sim myWorld = new WorldBankCovid19Sim(seed, p, "", demography);
 		return myWorld;
 	}
+	
 }
